Refactor code, remove unnecessary abstractions
